requirements.txt
peewee
beefish
boto
aaargh
grandfatherson
PyYaml
byteformat
pycrypto
requests
sh
events